+++ packages/components/src/components/dropdown/dropdown.tsx
@@ -2,7 +2,6 @@
 import { PropertyValues } from "lit";
 import { createEvent, h, JsxNode, LitElement, method, property, state } from "@arcgis/lumina";
 import { useDirection } from "@arcgis/lumina/controllers";
-import { nil } from "@arcgis/toolkit/type";
 import { nextFrame } from "../../utils/dom";
 import {
   connectFloatingUI,
@@ -15,6 +14,7 @@ import {
   hideFloatingUI,
   LogicalPlacement,
   OverlayPositioning,
+  ReferenceElement,
   reposition,
 } from "../../utils/floating-ui";
 import { isActivationKey } from "../../utils/key";
@@ -28,6 +28,12 @@ import type { DropdownGroup } from "../dropdown-group/dropdown-group";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import { useInteractive } from "../../controllers/useInteractive";
 import { useTopLayer } from "../../controllers/useTopLayer";
+import {
+  ReferenceElementComponent,
+  ReferenceElementType,
+  useReferenceElement,
+} from "../../controllers/useReferenceElement";
+import { referenceElementManager } from "../../controllers/useReferenceElement/manager";
 import { CSS, SLOTS } from "./resources";
 import { styles } from "./dropdown.scss";
 
@@ -37,11 +43,13 @@ declare global {
   }
 }
 
+const manager = referenceElementManager({ click: true, hover: true });
+
 /**
  * @slot - A slot for adding `calcite-dropdown-group` elements. Every `calcite-dropdown-item` must have a parent `calcite-dropdown-group`, even if the `groupTitle` property is not set.
- * @slot trigger - A slot for the element that triggers the component.
+ * @slot trigger - [deprecated] Use `referenceElement` property instead. A slot for the element that triggers the component.
  */
-export class Dropdown extends LitElement implements FloatingUIComponent {
+export class Dropdown extends LitElement implements FloatingUIComponent, ReferenceElementComponent {
   //#region Static Members
 
   static override shadowRootOptions = { mode: "open" as const, delegatesFocus: true };
@@ -51,6 +59,12 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   //#endregion
 
   //#region Private Properties
+
+  get referenceElementType(): ReferenceElementType | null {
+    return this.referenceElement ? this.type : null;
+  }
+
+  referenceElementController = useReferenceElement({ manager })(this);
 
   private direction = useDirection();
 
@@ -70,8 +84,6 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
 
   transitionProp = "opacity" as const;
 
-  referenceEl: HTMLDivElement;
-
   private resizeObserver = createObserver("resize", (entries) =>
     this.resizeObserverCallback(entries),
   );
@@ -79,6 +91,8 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   private scrollerEl: HTMLDivElement;
 
   transitionEl: HTMLDivElement;
+
+  onReferenceElementKeyDown = (event: KeyboardEvent): void => this.keyDownHandler(event);
 
   private focusSetter = useSetFocus<this>()(this);
 
@@ -93,6 +107,8 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   //#region State Properties
 
   @state() activeDescendantElement?: DropdownItem["el"];
+
+  @state() referenceEl: ReferenceElement;
 
   //#endregion
 
@@ -142,6 +158,17 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
    */
   @property({ reflect: true }) placement: LogicalPlacement = defaultMenuPlacement;
 
+  /**
+   * The `referenceElement` is used to position the component according to its `placement` value.
+   *
+   * Setting the value to an `HTMLElement` is preferred so the component does not need to query the DOM.
+   *
+   * However, a string `id` of the reference element can also be used.
+   *
+   *The component should not be placed within its own `referenceElement` to avoid unintended behavior.
+   */
+  @property() referenceElement: ReferenceElement | string | null;
+
   /** Specifies the size of the component. */
   @property({ reflect: true }) scale: Scale = "m";
 
@@ -157,7 +184,7 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
    *
    * Only set this if you need complex z-index control or if top layer placement causes conflicts with third-party components.
    *
-   * @mdn [Top Layer](https://developer.mozilla.org/en-US/docs/Glossary/Top_layer)
+   * @see [MDN - Top Layer](https://developer.mozilla.org/en-US/docs/Glossary/Top_layer)
    */
   @property({ reflect: true }) topLayerDisabled = false;
 
@@ -217,11 +244,14 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
    *
    * @param options - When specified an optional object customizes the component's focusing process. When `preventScroll` is `true`, scrolling will not occur on the component.
    *
-   * @mdn [focus(options)](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#options)
+   * @see [MDN - focus(options)](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#options)
    */
   @method()
   async setFocus(options?: FocusOptions): Promise<void> {
-    return this.focusSetter(() => this.referenceEl, options);
+    return this.focusSetter(
+      () => (this.referenceEl instanceof HTMLElement ? this.referenceEl : this.floatingEl),
+      options,
+    );
   }
 
   //#endregion
@@ -297,6 +327,16 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
     if (changes.has("scale") && (this.hasUpdated || this.scale !== "m")) {
       this.handlePropsChange();
     }
+
+    if (changes.has("referenceElement") && !this.referenceElement && this.open) {
+      this.topLayer.hide();
+    }
+  }
+
+  override updated(changes: PropertyValues<this>): void {
+    if (changes.has("referenceEl") && this.referenceElementType) {
+      connectFloatingUI(this);
+    }
   }
 
   loaded(): void {
@@ -340,7 +380,12 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   }
 
   private closeCalciteDropdownOnClick(event: MouseEvent): void {
-    if (this.disabled || !this.open || event.composedPath().includes(this.el)) {
+    if (
+      this.referenceElementType ||
+      this.disabled ||
+      !this.open ||
+      event.composedPath().includes(this.el)
+    ) {
       return;
     }
 
@@ -348,7 +393,7 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   }
 
   private closeCalciteDropdownOnOpenEvent(event: Event): void {
-    if (event.composedPath().includes(this.el)) {
+    if (this.referenceElementType || event.composedPath().includes(this.el)) {
       return;
     }
 
@@ -356,7 +401,7 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   }
 
   private pointerEnterHandler(): void {
-    if (this.disabled || this.type !== "hover") {
+    if (this.referenceElementType || this.disabled || this.type !== "hover") {
       return;
     }
 
@@ -364,7 +409,7 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   }
 
   private pointerLeaveHandler(): void {
-    if (this.disabled || this.type !== "hover") {
+    if (this.referenceElementType || this.disabled || this.type !== "hover") {
       return;
     }
 
@@ -438,7 +483,7 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   private setDropdownWidth(): void {
     const { referenceEl, scrollerEl } = this;
 
-    if (!scrollerEl || !referenceEl) {
+    if (!scrollerEl || !(referenceEl instanceof HTMLElement)) {
       return;
     }
 
@@ -487,8 +532,13 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   }
 
   private setReferenceEl(el: HTMLDivElement): void {
-    updateRefObserver(this.resizeObserver, this.referenceEl, el);
+    const previousReferenceEl = this.referenceEl instanceof HTMLElement ? this.referenceEl : null;
+    const nextReferenceEl = el instanceof HTMLElement ? el : null;
+
+    updateRefObserver(this.resizeObserver, previousReferenceEl, nextReferenceEl);
+
     this.referenceEl = el;
+
     connectFloatingUI(this);
   }
 
@@ -498,7 +548,10 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
   }
 
   private keyDownHandler(event: KeyboardEvent): void {
-    if (!event.composedPath().includes(this.referenceEl)) {
+    if (
+      !(this.referenceEl instanceof HTMLElement) ||
+      !event.composedPath().includes(this.referenceEl)
+    ) {
       return;
     }
 
@@ -623,7 +676,7 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
     this.updateActiveDescendantElement(activeItem);
   }
 
-  private updateActiveDescendantElement(activeItem: DropdownItem["el"] | nil): void {
+  private updateActiveDescendantElement(activeItem: DropdownItem["el"] | null): void {
     this.items.forEach((item) => {
       item.activeDescendant = item === activeItem;
     });
@@ -700,7 +753,9 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
     if (
       relatedTarget &&
       (this.el.contains(relatedTarget) ||
-        (this.referenceEl != null && this.referenceEl.contains(relatedTarget)))
+        (this.referenceEl != null &&
+          this.referenceEl instanceof HTMLElement &&
+          this.referenceEl.contains(relatedTarget)))
     ) {
       return;
     }
@@ -724,22 +779,24 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
     const { open } = this;
     return (
       <this.interactiveContainer disabled={this.disabled}>
-        <div
-          class={CSS.triggerContainer}
-          onClick={this.toggleClickDropdown}
-          onFocusIn={this.openHoverDropdown}
-          onFocusOut={this.closeHoverDropdown}
-          onKeyDown={this.keyDownHandler}
-          ref={this.setReferenceEl}
-        >
-          <slot
-            ariaActiveDescendantElement={this.activeDescendantElement ?? null}
-            ariaControlsElements={this.scrollerEl ? [this.scrollerEl] : undefined}
-            ariaExpanded={open}
-            ariaHasPopup="menu"
-            name={SLOTS.trigger}
-          />
-        </div>
+        {!this.referenceElementType ? (
+          <div
+            class={CSS.triggerContainer}
+            onClick={this.toggleClickDropdown}
+            onFocusIn={this.openHoverDropdown}
+            onFocusOut={this.closeHoverDropdown}
+            onKeyDown={this.keyDownHandler}
+            ref={this.setReferenceEl}
+          >
+            <slot
+              ariaActiveDescendantElement={this.activeDescendantElement ?? null}
+              ariaControlsElements={this.scrollerEl ? [this.scrollerEl] : undefined}
+              ariaExpanded={open}
+              ariaHasPopup="menu"
+              name={SLOTS.trigger}
+            />
+          </div>
+        ) : null}
         <div
           class={{
             [CSS.wrapper]: true,
@@ -752,7 +809,9 @@ export class Dropdown extends LitElement implements FloatingUIComponent {
           ref={this.setFloatingEl}
         >
           <div
-            ariaLabelledByElements={this.referenceEl ? [this.referenceEl] : undefined}
+            ariaLabelledByElements={
+              this.referenceEl instanceof HTMLElement ? [this.referenceEl] : undefined
+            }
             class={{
               [CSS.content]: true,
               [FloatingCSS.animation]: true,